Move InlineContent into InlineSequence. Adjust types.
