+++ .gitignore
@@ -1,6 +1,5 @@
 .idea
 node_modules
-/types
 *.log
 .eslintcache
 .DS_Store
