+++ src/blocks/Inputs.tsx
@@ -16,6 +16,11 @@ export const inputsProps = z.object({
 
 export type InputsProps = z.infer<typeof inputsProps>
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isInputs(o: any): o is InputsProps {
+  return 'inputs' in o
+}
+
 export const Inputs = (props: InputsProps) => (
   <Placeholder {...inputsProps.parse(props)} label="Inputs block" />
 )
