+++ src/inlines/InlineSequence.tsx
@@ -1,9 +1,14 @@
 import { z } from 'zod'
-import { inlineSequence } from '../inlines'
-import { Text, isTextProps } from '../inlines/Text'
-import { Icon, isIconProps } from '../inlines/Icon'
-import { invalidInline } from '../lib/warnings'
 import { key } from '../lib/keys'
+import { Text, textProps, isTextProps } from './Text'
+import { Icon, iconProps, isIconProps } from './Icon'
+import { invalidInline } from '../lib/warnings'
+
+export const inlineProps = z.union([textProps, iconProps])
+
+export const inlineSequence = z.array(inlineProps)
+
+export type InlineSequence = z.infer<typeof inlineSequence>
 
 export const inlineContentProps = z.object({
   inlines: inlineSequence,
@@ -19,13 +24,15 @@ export const InlineContent = (props: InlineContentProps) => {
   const { inlines } = inlineContentProps.parse(props)
   return (
     <>
-      {(inlines ?? []).map((inline) => isIconProps(inline) ? (
+      {(inlines ?? []).map((inline) =>
+        isIconProps(inline) ? (
           <Icon {...inline} key={key(inline)} />
         ) : isTextProps(inline) ? (
           <Text {...inline} key={key(inline)} />
         ) : (
           invalidInline(inline)
-        ))}
+        )
+      )}
     </>
   )
 }
