+++ src/inputs/radio-group.ts
@@ -1,9 +1,10 @@
 import { z } from 'zod'
 import { inputPropsWithInitialStringValue } from './input-properties'
+import { inlineSequence } from '../inlines'
 
 export const radioItemProps = z.object({
   value: z.string().nonempty(),
-  label: z.string().nonempty(), // TODO: change to `phrasingContentProps`
+  label: inlineSequence,
 })
 
 export const radioGroupProps = inputPropsWithInitialStringValue.merge(
